requirements.txt
matplotlib==3.4.3
numpy==1.21.0
opencv_python==4.5.4.60
pandas==1.3.4
Pillow==8.4.0
scikit_learn==1.0.1
tensorboardX==2.4.1
torch==1.8.0
torchvision==0.10.0
tqdm==4.62.3
